+++ aws/fetch/manual_fetchers.go
@@ -1659,3 +1659,6 @@ func addManualMskFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
 
 func addManualMqFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
 }
+
+func addManualFsxFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
+}