+++ aws/services/gen_services.go
@@ -77,6 +77,8 @@ import (
 	elbv2types "github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2/types"
 	eventbridge "github.com/aws/aws-sdk-go-v2/service/eventbridge"
 	eventbridgetypes "github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
+	fsx "github.com/aws/aws-sdk-go-v2/service/fsx"
+	fsxtypes "github.com/aws/aws-sdk-go-v2/service/fsx/types"
 	glue "github.com/aws/aws-sdk-go-v2/service/glue"
 	gluetypes "github.com/aws/aws-sdk-go-v2/service/glue/types"
 	iam "github.com/aws/aws-sdk-go-v2/service/iam"
@@ -151,6 +153,7 @@ var ServiceNames = []string{
 	"kinesis",
 	"redshift",
 	"codepipeline",
+	"fsx",
 	"mq",
 	"msk",
 	"cognito",
@@ -242,6 +245,8 @@ var ResourceTypes = []string{
 	"redshiftcluster",
 	"redshiftsubnetgroup",
 	"pipeline",
+	"fsxfilesystem",
+	"fsxbackup",
 	"broker",
 	"kafkacluster",
 	"userpool",
@@ -296,6 +301,7 @@ var ServicePerAPI = map[string]string{
 	"kinesis":                 "kinesis",
 	"redshift":                "redshift",
 	"codepipeline":            "codepipeline",
+	"fsx":                     "fsx",
 	"mq":                      "mq",
 	"kafka":                   "msk",
 	"cognitoidentityprovider": "cognito",
@@ -388,6 +394,8 @@ var ServicePerResourceType = map[string]string{
 	"redshiftcluster":          "redshift",
 	"redshiftsubnetgroup":      "redshift",
 	"pipeline":                 "codepipeline",
+	"fsxfilesystem":            "fsx",
+	"fsxbackup":                "fsx",
 	"broker":                   "mq",
 	"kafkacluster":             "msk",
 	"userpool":                 "cognito",
@@ -486,6 +494,8 @@ var APIPerResourceType = map[string]string{
 	"redshiftcluster":          "redshift",
 	"redshiftsubnetgroup":      "redshift",
 	"pipeline":                 "codepipeline",
+	"fsxfilesystem":            "fsx",
+	"fsxbackup":                "fsx",
 	"broker":                   "mq",
 	"kafkacluster":             "kafka",
 	"userpool":                 "cognitoidentityprovider",
@@ -5010,6 +5020,157 @@ func (s *Codepipeline) FetchByType(ctx context.Context, t string) (cloud.GraphAP
 
 func (s *Codepipeline) IsSyncDisabled() bool {
 	return !getBool(s.config, "aws.codepipeline.sync", true)
+}
+
+type Fsx struct {
+	fetcher         fetch.Fetcher
+	region, profile string
+	config          map[string]any
+	log             *logger.Logger
+	FsxClient       *fsx.Client
+}
+
+func NewFsx(cfg aws.Config, profile string, extraConf map[string]any, log *logger.Logger) cloud.Service {
+	region := cfg.Region
+	fsxClient := fsx.NewFromConfig(cfg)
+
+	fetchConfig := awsfetch.NewConfig(
+		fsxClient,
+	)
+	fetchConfig.Extra = extraConf
+	fetchConfig.Log = log
+
+	return &Fsx{
+		FsxClient: fsxClient,
+		fetcher:   fetch.NewFetcher(awsfetch.BuildFsxFetchFuncs(fetchConfig)),
+		config:    extraConf,
+		region:    region,
+		profile:   profile,
+		log:       log,
+	}
+}
+
+func (s *Fsx) Name() string {
+	return "fsx"
+}
+
+func (s *Fsx) Region() string {
+	return s.region
+}
+
+func (s *Fsx) Profile() string {
+	return s.profile
+}
+
+func (s *Fsx) ResourceTypes() []string {
+	return []string{
+		"fsxfilesystem",
+		"fsxbackup",
+	}
+}
+
+func (s *Fsx) Fetch(ctx context.Context) (cloud.GraphAPI, error) {
+	if s.IsSyncDisabled() {
+		return graph.NewGraph(), nil
+	}
+
+	allErrors := new(fetch.Error)
+
+	gph, err := s.fetcher.Fetch(context.WithValue(ctx, "region", s.region))
+	defer s.fetcher.Reset()
+
+	for _, e := range *fetch.WrapError(err) {
+		switch ee := e.(type) {
+		case nil:
+			continue
+		default:
+			var ae smithy.APIError
+			if errors.As(ee, &ae) && ae.ErrorMessage() == accessDenied {
+				allErrors.Add(cloud.ErrFetchAccessDenied)
+			} else {
+				allErrors.Add(ee)
+			}
+		}
+	}
+
+	if err := gph.AddResource(graph.InitResource(cloud.Region, s.region)); err != nil {
+		return gph, err
+	}
+
+	snap := gph.AsRDFGraphSnaphot()
+
+	errc := make(chan error)
+	var wg sync.WaitGroup
+	if getBool(s.config, "aws.fsx.fsxfilesystem.sync", true) {
+		list, err := s.fetcher.Get("fsxfilesystem_objects")
+		if err != nil {
+			return gph, err
+		}
+		if _, ok := list.([]fsxtypes.FileSystem); !ok {
+			return gph, errors.New("cannot cast to '[]fsxtypes.FileSystem' type from fetch context")
+		}
+		for _, r := range list.([]fsxtypes.FileSystem) {
+			for _, fn := range addParentsFns["fsxfilesystem"] {
+				wg.Add(1)
+				go func(f addParentFn, snap tstore.RDFGraph, region string, res *fsxtypes.FileSystem) {
+					defer wg.Done()
+					err := f(gph, snap, region, res)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}(fn, snap, s.region, &r)
+			}
+		}
+	}
+	if getBool(s.config, "aws.fsx.fsxbackup.sync", true) {
+		list, err := s.fetcher.Get("fsxbackup_objects")
+		if err != nil {
+			return gph, err
+		}
+		if _, ok := list.([]fsxtypes.Backup); !ok {
+			return gph, errors.New("cannot cast to '[]fsxtypes.Backup' type from fetch context")
+		}
+		for _, r := range list.([]fsxtypes.Backup) {
+			for _, fn := range addParentsFns["fsxbackup"] {
+				wg.Add(1)
+				go func(f addParentFn, snap tstore.RDFGraph, region string, res *fsxtypes.Backup) {
+					defer wg.Done()
+					err := f(gph, snap, region, res)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}(fn, snap, s.region, &r)
+			}
+		}
+	}
+
+	go func() {
+		wg.Wait()
+		close(errc)
+	}()
+
+	for err := range errc {
+		if err != nil {
+			allErrors.Add(err)
+		}
+	}
+
+	if allErrors.Any() {
+		return gph, allErrors
+	}
+
+	return gph, nil
+}
+
+func (s *Fsx) FetchByType(ctx context.Context, t string) (cloud.GraphAPI, error) {
+	defer s.fetcher.Reset()
+	return s.fetcher.FetchByType(context.WithValue(ctx, "region", s.region), t)
+}
+
+func (s *Fsx) IsSyncDisabled() bool {
+	return !getBool(s.config, "aws.fsx.sync", true)
 }
 
 type Mq struct {