+++ aws/services/init.go
@@ -30,7 +30,7 @@ var (
 	AccessService, InfraService, StorageService, MessagingService, DNSService, LambdaService, MonitoringService, CdnService, CloudformationService       cloud.Service
 	EksService, DynamodbService, SecretsmanagerService, ApigatewayService, SsmService, EfsService, CloudtrailService, CloudwatchlogsService              cloud.Service
 	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService, KinesisService, RedshiftService, CodepipelineService cloud.Service
-	CodebuildService, BeanstalkService, CodedeployService, GlueService, SesService, CognitoService, MskService, MqService                                cloud.Service
+	CodebuildService, BeanstalkService, CodedeployService, GlueService, SesService, CognitoService, MskService, MqService, FsxService                    cloud.Service
 )
 
 func Init(profile, region string, extraConf map[string]any, log *logger.Logger, profileSetterCallback func(val string) error, enableNetworkMonitor bool) error {
@@ -79,6 +79,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	CognitoService = NewCognito(cfg, profile, extraConf, log)
 	MskService = NewMsk(cfg, profile, extraConf, log)
 	MqService = NewMq(cfg, profile, extraConf, log)
+	FsxService = NewFsx(cfg, profile, extraConf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -113,6 +114,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	cloud.ServiceRegistry[CognitoService.Name()] = CognitoService
 	cloud.ServiceRegistry[MskService.Name()] = MskService
 	cloud.ServiceRegistry[MqService.Name()] = MqService
+	cloud.ServiceRegistry[FsxService.Name()] = FsxService
 
 	awsspec.CommandFactory = &awsspec.AWSFactory{
 		Log: log,